SamplePlayer.js → SamplePlayer.svelte
Okay, this bit's easy -- next comes the more difficult part...
